requirements.txt

python-telegram-bot>=20.7,<21
python-dotenv>=1.0.1
ccxt>=4.4.0
pandas>=2.2.2
numpy>=1.26.4
tenacity>=8.4.2
